[MOVE/#403] 상점 배너 폴더링 분리
