+++ app/src/main/java/com/el/yello/presentation/pay/banner/PayBannerViewHolder.kt
@@ -1,0 +1,24 @@
+package com.el.yello.presentation.pay.banner
+
+import androidx.recyclerview.widget.RecyclerView
+import com.el.yello.databinding.ItemPayFirstBinding
+import com.el.yello.databinding.ItemPaySecondBinding
+import com.el.yello.databinding.ItemPayThirdBinding
+
+class PayOneViewHolder(
+    binding: ItemPayFirstBinding,
+) : RecyclerView.ViewHolder(binding.root) {
+    fun onBind() {}
+}
+
+class PayTwoViewHolder(
+    binding: ItemPaySecondBinding,
+) : RecyclerView.ViewHolder(binding.root) {
+    fun onBind() {}
+}
+
+class PayThreeViewHolder(
+    binding: ItemPayThirdBinding,
+) : RecyclerView.ViewHolder(binding.root) {
+    fun onBind() {}
+}
